move clients to new subdir
