+++ lib/clients/__mocks__/websocket-client.js
