+++ lib/clients/facebook-messenger-client.js
@@ -1,4 +1,4 @@
-var Session = require('./session')
+var Session = require('../session')
 var request = require('request')
 
 function FacebookMessengerClient(config) {
