+++ lib/clients/index.js
@@ -1,0 +1,11 @@
+const FacebookMessengerClient = require('./facebook-messenger-client')
+const TwilioClient = require('./twilio-client')
+const TwitterClient = require('./twitter-client')
+const WebsocketClient = require('./websocket-client')
+
+module.exports = {
+  FacebookMessengerClient,
+  TwilioClient,
+  TwitterClient,
+  WebsocketClient,
+}
