+++ lib/clients/twilio-client.js
@@ -1,4 +1,4 @@
-var Session = require('./session')
+var Session = require('../session')
 
 // Currently this client only supports texts but in the future could support calls
 function TwilioClient(config) {
