+++ lib/clients/twitter-client.js
@@ -1,4 +1,4 @@
-var Session = require('./session')
+var Session = require('../session')
 var Twit = require('twit')
 
 function TwitterClient(config) {
