+++ lib/clients/websocket-client.js
@@ -1,4 +1,4 @@
-var Session = require('./session')
+var Session = require('../session')
 
 function WebsocketClient(io) {
   return function(bot) {
