+++ lib/index.js
@@ -1,16 +1,22 @@
-var Bot = require('./bot')
-var FacebookMessengerClient = require('./facebook-messenger-client')
-var Server = require('./server')
-var TwilioClient = require('./twilio-client')
-var TwitterClient = require('./twitter-client')
-var WebsocketClient = require('./websocket-client')
+const Bot = require('./bot')
+const Clients = require('./clients')
+const Server = require('./server')
+
+const FacebookMessengerClient = Clients.FacebookMessengerClient
+const TwilioClient = Clients.TwilioClient
+const TwitterClient = Clients.TwitterClient
+const WebsocketClient = Clients.WebsocketClient
 
 require('./listen')
 
 module.exports = {
-  Bot: Bot,
-  FacebookMessengerClient: FacebookMessengerClient,
-  Server: Server,
-  TwilioClient: TwilioClient,
-  TwitterClient: TwitterClient
+  Bot,
+  Clients,
+  Server,
+
+  // Clients
+  FacebookMessengerClient,
+  TwilioClient,
+  TwitterClient,
+  WebsocketClient,
 }
