+++ lib/server.js
@@ -1,5 +1,5 @@
 var Bot = require('./bot')
-var WebsocketClient = require('./websocket-client')
+var WebsocketClient = require('./clients/websocket-client')
 
 function Server() {
   this.namespaces = {}
